feat: Update rename file index - main menu - index Update data reload in localstorage and backup data when user reload the page.
