+++ admin/index.html
@@ -12,7 +12,7 @@
 <div class="container">
     <div class="row mt-3 mb-3">
         <div class="col-2 d-flex align-items-center">
-            <a href="../index.html" class="link-underline link-underline-opacity-0">
+            <a href="../main.html" class="link-underline link-underline-opacity-0">
                 <button type="button" class="btn btn-outline-secondary">Quay lại</button>
             </a>
         </div>
