+++ main.html
@@ -1,0 +1,123 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Hangman Game</title>
+    <!-- Font-Awesome -->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+    <!-- Bootstrap -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
+    <link rel="stylesheet" href="main.css">
+    <script src="js/time.js" defer></script>
+    <script type="module" src="js/main.js"></script>
+</head>
+<body>
+    <div class="game-modal">
+        <div class="content">
+            <img src="img/lost.gif" alt="gif">
+            <h4>"Result!"</h4>
+            <p>Bạn đã tìm thấy từ: <b>"key word"</b></p>
+            <button class="play-again">Chơi lại</button>
+        </div>
+    </div>
+    <div class="container-main">
+        <div class="wrapper">
+            <button type="button" id = "list-score" class="btn btn-result flex-fill fw-bold p-2" data-bs-toggle="modal" data-bs-target="#resultModal">
+                <i class="fas fa-list"></i>
+            </button>
+            <div class="container-info">
+                <h5 class="bungee-regular-title" id="team-name">*****</h5>
+                <h5 class="bungee-regular-title text-action" id="title"><i class="fas fa-hourglass-end"></i> <span class="bungee-regular-title" id="timer">03:00</span></h5>
+            </div>
+            <div class="container">
+                <div class="hangman-box">
+                    <img src="img/hangman-0.svg" alt="hangman-img">
+                    <h1 class="press-start-2p-regular">Hangman Game</h1>
+                </div>
+                <div class="game-box">
+                    <ul class="word-display">
+                        <!-- <li class="letter"></li>
+                        <li class="letter"></li>
+                        <li class="letter guessed">a</li>
+                        <li class="letter"></li>
+                        <li class="letter"></li>
+                        <li class="letter"></li>
+                        <li class="letter"></li> -->
+                    </ul>
+                    <h4 class="hint-text ">
+                        Gợi ý: <b>"Hint"</b>
+                    </h4>
+                    <h4 class="guesses-text">
+                        Suy đoán không chính xác: <b>0 / 6</b>
+                    </h4>
+                    <div class="keyboard bungee-regular-normal">
+                        <!-- implement button keyboard-->
+                    </div>
+                </div>
+            </div>
+        </div>
+        <div class="list-core">
+            <table class="table">
+                <thead>
+                <tr>
+                    <th scope="col"><img width="32" height="32" src="https://img.icons8.com/pulsar-color/48/rank.png" alt="rank"/></th>
+                    <th class="bungee-regular-purple" scope="col">Tên</th>
+                    <th class="bungee-regular-purple" scope="col">Điểm</th>
+                </tr>
+                </thead>
+                <tbody>
+                <!-- This will be populated by updateLeaderboard -->
+                </tbody>
+            </table>
+        </div>
+    </div>
+    <!-- Modal -->
+    <div class="modal fade" id="resultModal" tabindex="-1" aria-labelledby="resultModalLabel" aria-hidden="true">
+        <div class="modal-dialog modal-lg modal-dialog-centered modal-dialog-scrollable">
+            <div class="modal-content">
+                <div class="modal-header border-0">
+                    <h5 class="modal-title text-center w-100 bungee-regular-normal-purple" id="resultModalLabel">Danh sách điểm</h5>
+                </div>
+                <div class="modal-body">
+                    <table class="table">
+                        <thead>
+                        <tr>
+                            <th scope="col"><img width="32" height="32" src="https://img.icons8.com/pulsar-color/48/rank.png" alt="rank"/></th>
+                            <th class="bungee-regular-purple" scope="col">Tên</th>
+                            <th class="bungee-regular-purple" scope="col">Điểm</th>
+                        </tr>
+                        </thead>
+                        <tbody>
+                        <!-- This will be populated by updateLeaderboard -->
+                        </tbody>
+                    </table>
+                </div>
+                <div class="modal-footer border-0 d-flex justify-content-center">
+                    <button type="button" class="btn btn-secondary text-action btn-action" data-bs-dismiss="modal">Đóng</button>
+                </div>
+            </div>
+        </div>
+    </div>
+<!--    Modal Alert Game-->
+    <div class="modal fade" id="alertModal" tabindex="-1" aria-labelledby="alertModalLabel" data-bs-target="static" data-bs-keyboard="false" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered modal-xs">
+            <div class="modal-content">
+                <div class="modal-header border-0">
+                    <h5 class="modal-title text-center w-100 bungee-regular-normal-purple" id="alertModalLabel">Kết thúc lượt chơi</h5>
+                </div>
+                <div class="modal-body d-flex justify-content-center">
+                    <div class="body-content d-flex justify-content-center flex-column w-50"></div>
+                </div>
+                <div class="modal-footer border-0 d-flex justify-content-center">
+                    <button type="button" id="confirmBtn" class="btn btn-primary text-action btn-action">Xác nhận</button>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+
+</body>
+</html>
