--- menu.html
@@ -1,112 +1,0 @@
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Menu Hangman</title>
-    <!-- Font-Awesome -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
-    <!-- Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <!-- CSS -->
-    <link rel="stylesheet" href="menu.css">
-</head>
-<body id="body">
-<div class="wrapper">
-    <div class="align" id="align">
-        <h4 id="title" class="title-game press-start-2p-regular text-shadow">******* ****</h4>
-    </div>
-    <div class="container">
-            <div class="row-12 w-100">
-                <ul class="threeD-button-set">
-                    <li>
-                        <div class="col m-3 d-flex justify-content-center align-items-center hover">
-                            <a href="index.html" type="button" class="btn btn-primary flex-fill fw-bold p-2" data-bs-toggle="modal" data-bs-target="#configurationModal">
-                                PLAY
-                            </a>
-                        </div>
-                    </li>
-<!--                    <li>-->
-<!--                        <div class="col m-3 d-flex justify-content-center align-items-center">-->
-<!--                            <button type="button" class="btn btn-primary flex-fill fw-bold p-2" data-bs-toggle="modal" data-bs-target="#configurationModal">-->
-<!--                                OPTIONS-->
-<!--                            </button>-->
-<!--                        </div>-->
-<!--                    </li>-->
-                    <li>
-                        <div class="col m-3 d-flex justify-content-center align-items-center">
-                            <button type="button" class="btn btn-primary flex-fill fw-bold p-2" data-bs-toggle="modal" data-bs-target="#guideModal">
-                                GUIDE
-                            </button>
-                        </div>
-                    </li>
-                </ul>
-            </div>
-    </div>
-</div>
-  <!-- Configuration Modal -->
-<div class="modal fade" id="configurationModal" tabindex="-1" aria-labelledby="configurationModalLabel" aria-hidden="true">
-    <div class="modal-dialog modal-dialog-centered modal-xs">
-      <div class="modal-content">
-        <div class="modal-header border-0">
-            <button type="button" class="btn-closed d-none" id="backBtn">
-                <img width="32" height="32" src="https://img.icons8.com/pieces/64/circled-chevron-left.png" alt="circled-chevron-left"/>
-            </button>
-          <h5 class="modal-title text-center w-100 bungee-regular-title" id="configurationModalLabel">Cấu hình Game</h5>
-          <button type="button" class="btn-closed" data-bs-dismiss="modal" aria-label="Close"><img width="32" height="32" src="https://img.icons8.com/pieces/64/x-coordinate.png" alt="x-coordinate"/></button>
-        </div>
-        <div class="modal-body">
-          <form action="#" novalidate>
-            <div id="page-one" class="page">
-                <div class="form-group m-3">
-                    <label for="teamNumber">Số đội chơi</label>
-                    <input type="number" class="form-control" id="teamNumber" placeholder="Nhập số đội chơi" min="1" max="5" required>
-                    <div class="error" id="teamNumberError"></div>
-                </div>
-                <div class="form-group m-3">
-                    <label for="limitedTime">Giới hạn thời gian</label>
-                    <input type="text" class="form-control" id="limitedTime" placeholder="Nhập thời gian chơi mỗi vòng" required>
-                    <div class="error" id="limitedTimeError"></div>
-                </div>
-                <div class="form-group m-3">
-                    <label for="limitedReasoning">Giới hạn số lần suy đoán</label>
-                    <input type="text" class="form-control" id="limitedReasoning" placeholder="Nhập giới hạn suy đoán sai cho mỗi vòng" required>
-                    <div class="error" id="limitedReasoningError"></div>
-                </div>
-            </div>
-            <div id="page-two" class="page d-none">
-                <div id="teamNameFields" class="overflow-auto"></div>
-            </div>
-          </form>
-        </div>
-        <div class="modal-footer border-0 d-flex justify-content-center">
-          <button type="button" class="btn btn-primary text-action btn-action" id="nextBtn">Tiếp theo</button>
-          <button type="button" class="btn btn-primary text-action btn-action d-none" id="startBtn">Bắt đầu</button>
-        </div>
-      </div>
-    </div>
-</div>
-<!-- Guide Modal -->
-<div class="modal fade" id="guideModal" tabindex="-1" aria-labelledby="guideModalLabel" aria-hidden="true">
-    <div class="modal-dialog modal-lg modal-dialog-centered modal-dialog-scrollable">
-      <div class="modal-content">
-        <div class="modal-header border-0">
-          <h5 class="modal-title text-center w-100 bungee-regular-title" id="guideModalLabel">Hướng dẫn chơi Game</h5>
-          <button type="button" class="btn-closed" data-bs-dismiss="modal" aria-label="Close">
-              <img width="32" height="32" src="https://img.icons8.com/pieces/64/circled-chevron-left.png" alt="circled-chevron-left"/>
-          </button>
-        </div>
-        <div class="modal-body">
-          <p class="content"></p>
-        </div>
-        <div class="modal-footer border-0 d-flex justify-content-center">
-            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Đóng</button>
-        </div>
-      </div>
-    </div>
-</div>
-<script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-
-<script type="module" src="js/menu.js"></script>
-</body>
-</html>
